backend/leo-kos-backend/app/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.17.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
motor==3.0.0
passlib[bcrypt]==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pymongo[srv]==4.1.1
python-dotenv==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
